restore: hide db operations behind interface
