+++ pkg/service/backup/restore_progress.go
@@ -5,9 +5,6 @@ package backup
 import (
 	"context"
 	"strings"
-
-	"github.com/scylladb/gocqlx/v2/qb"
-	"github.com/scylladb/scylla-manager/v3/pkg/schema/table"
 )
 
 // aggregateProgress returns restore progress information classified by keyspace and tables.
@@ -38,7 +35,7 @@ func (w *restoreWorkerTools) aggregateProgress(ctx context.Context, run *Restore
 	}
 
 	// Initialize tables' progress
-	w.ForEachProgress(ctx, run, aggregateRestoreTableProgress(tableMap))
+	w.cacheProvider.ForEachRestoreProgress(ctx, run, aggregateRestoreTableProgress(tableMap))
 
 	// Aggregate progress
 	for _, u := range run.Units {
@@ -121,68 +118,4 @@ func (rp *restoreProgress) calcParentProgress(child restoreProgress) {
 
 	rp.StartedAt = calcParentStartedAt(rp.StartedAt, child.StartedAt)
 	rp.CompletedAt = calcParentCompletedAt(rp.CompletedAt, child.CompletedAt)
-}
-
-// ForEachProgress iterates over all RestoreRunProgress that belong to the run.
-// NOTE: callback is always called with the same pointer - only the value that it points to changes.
-func (w *restoreWorkerTools) ForEachProgress(ctx context.Context, run *RestoreRun, cb func(*RestoreRunProgress)) {
-	iter := table.RestoreRunProgress.SelectQuery(w.managerSession).BindMap(qb.M{
-		"cluster_id": run.ClusterID,
-		"task_id":    run.TaskID,
-		"run_id":     run.ID,
-	}).Iter()
-	defer func() {
-		if err := iter.Close(); err != nil {
-			w.Logger.Error(ctx, "Error while iterating over run progress",
-				"cluster_id", run.ClusterID,
-				"task_id", run.TaskID,
-				"run_id", run.ID,
-				"error", err,
-			)
-		}
-	}()
-
-	pr := new(RestoreRunProgress)
-	for iter.StructScan(pr) {
-		cb(pr)
-	}
-}
-
-// ForEachTableProgress iterates over all RestoreRunProgress that belong to the run
-// with the same manifest, keyspace and table as the run.
-// NOTE: callback is always called with the same pointer - only the value that it points to changes.
-func (w *restoreWorkerTools) ForEachTableProgress(ctx context.Context, run *RestoreRun, cb func(*RestoreRunProgress)) {
-	iter := qb.Select(table.RestoreRunProgress.Name()).Where(
-		qb.Eq("cluster_id"),
-		qb.Eq("task_id"),
-		qb.Eq("run_id"),
-		qb.Eq("manifest_path"),
-		qb.Eq("keyspace_name"),
-		qb.Eq("table_name"),
-	).Query(w.managerSession).BindMap(qb.M{
-		"cluster_id":    run.ClusterID,
-		"task_id":       run.TaskID,
-		"run_id":        run.ID,
-		"manifest_path": run.ManifestPath,
-		"keyspace_name": run.Keyspace,
-		"table_name":    run.Table,
-	}).Iter()
-	defer func() {
-		if err := iter.Close(); err != nil {
-			w.Logger.Error(ctx, "Error while iterating over table's run progress",
-				"cluster_id", run.ClusterID,
-				"task_id", run.TaskID,
-				"run_id", run.ID,
-				"manifest_path", run.ManifestPath,
-				"keyspace", run.Keyspace,
-				"table", run.Table,
-				"error", err,
-			)
-		}
-	}()
-
-	pr := new(RestoreRunProgress)
-	for iter.StructScan(pr) {
-		cb(pr)
-	}
 }
